requirements.txt
discord.py
python-dotenv
openai 
numpy 
scikit-learn
loguru
beautifulsoup4
pymongo